scripts/data_access_migration_from_gh_to_aws/requirements.dev.txt
boto3
pygithub==1.34
